requirements.txt
asgiref==3.5.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
cycler==0.11.0
decorator==5.1.1
Django==4.0.4
fonttools==4.33.3
googlesearch-python==1.1.0
greenlet==1.1.2
idna==2.10
joblib==1.1.0
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.1.5
Pillow==9.1.1
py==1.11.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.25.1
retry==0.9.2
scikit-learn==1.1.0
scipy==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
sqlparse==0.4.2
threadpoolctl==3.1.0
urllib3==1.26.9
whois==0.9.13
yara-python==4.2.0
